feat: implement filesystem store
