+++ tests/checkpoint/processor_test.py
@@ -20,7 +20,7 @@ from aws_durable_execution_sdk_python_testing.exceptions import (
 )
 from aws_durable_execution_sdk_python_testing.execution import Execution
 from aws_durable_execution_sdk_python_testing.scheduler import Scheduler
-from aws_durable_execution_sdk_python_testing.stores import ExecutionStore
+from aws_durable_execution_sdk_python_testing.stores.base import ExecutionStore
 from aws_durable_execution_sdk_python_testing.token import CheckpointToken
